add API

+++ app/src/main/java/com/lepquold/model/Clothing.java
@@ -4,12 +4,12 @@ public class Clothing {
     private int clothingID;
     public String name;
     private String description;
-    public float temperature;
+    public double temperature;
     public boolean isWaterProof;
     public Style style;
     public Type type;
 
-    public Clothing(String description, float temperature, boolean isWaterProof, Style style, Type type) {
+    public Clothing(String description, double temperature, boolean isWaterProof, Style style, Type type) {
         this.description = description;
         this.temperature = temperature;
         this.isWaterProof = isWaterProof;
