+++ app/src/main/java/com/lepquold/model/OutfitRequest.java
@@ -1,7 +1,8 @@
 package com.lepquold.model;
 
 public class OutfitRequest {
-    public float temperature;
+    public double temperature;
     public boolean isRaining;
     public Style style;
+
 }
